Moved VirusTotal stuff to vt.go
